fix tree corruption upon delete then insert
This data corruption is due to the fact that entries of different nodes share the same slice.
For example, assume

nodeA.entries == buf[:3]
nodeB.entries == buf[3:]

when we do

nodeA.entries = append(nodeA.entries, foo)

we accidentally make nodeB.entries[0] == foo
and nodeB's data is corrupted.

This issue is likely the cause behind
https://github.com/dhconnelly/rtreego/issues/28