Reduce unnecessary allocations and sorting to improve performance
As shown in the pprof graph to be attached in the pull request,
the previous implementation has the following time spent:

gcBgMarkWorker 24.37%
mallocgc       15.21%
sort.Sort      9.06%
gcWriteBarrier 7.2%

which adds up to almost half the running time.

The pprof graph after this change (to be also attached) shows that these memory allocations
and sorting operations are indeed reduced.