make entry.bb plain Rect struct instead of pointer
This is to improve performance by avoiding garbage collection.
As shown in the pprof graphs to be uploaded in the pull request,
before this change the time spent in GC is:

runtime.mallocgc   18.19%
runtime.scanobject 25.23%

which adds up to almost half the running time.

After this change, the time spent is:

runtime.mallocgc   7.08%
runtime.scanobject 6.40%

which is a significant reduction.

This remaining time wasted in GC can be further shaved off if Point is
defined with fixed sized arrays, instead of slices.
However, fixed sized arrays run counter to the aspirations of this
library so it is not pursued here.