remove unneeeded allocation in computeBoundingBox
This improves performance by reducing GC for Insert and Delete intensive
workloads.
As seen in the pprof graph to be attached in the pull request,
before this change time was spent on:

computeBoundingBox  16.26%
runtime.systemstack 20.06%

which adds up to almost 40%.

After this change, computeBoundingBox spends only 2%, and all traces of
GC are gone.